Show friend requester name instead of id

+++ inc/friendScript.php
@@ -10,37 +10,63 @@ if(isset($_POST['type']) && $_POST['type'] != ""){
       $curUser = $_POST['curUser'];
       $othUser = $_POST['othUser'];
 
-      $stmnt = $container->database->prepare("INSERT INTO friend_request (current_user_id, other_user_id, friend_requester, request_time)
-                       VALUES (:curUs, :othUs, :friReq, :reqTim);");
-      $stmnt->execute([
-        'curUs' => $curUser,
-        'othUs' => $othUser,
-        'friReq' => $curUser,
-        'reqTim' => date("y-m-d H:i:s"),
-      ]);
-      echo "success";
+      $friendAddName = "unset";
+
+      /* Retrieve the name of the requester */
+      $stmnt = $container->database->prepare("SELECT account_name, account_id FROM accounts WHERE account_id = :accId");
+      $stmnt->execute(['accId' => $curUser]);
+      $result = $stmnt->fetchAll(PDO::FETCH_ASSOC);
+      if(count($result) > 0){
+        foreach($result as $data){
+          $friendAddName = $data['account_name'];
+        }
+      }
+      /* Insert a friend request into a table */
+      if($friendAddName != "unset"){
+        $stmnt = $container->database->prepare("INSERT INTO friend_request (current_user_id, other_user_id, friend_requester, request_time)
+        VALUES (:curUs, :othUs, :friReq, :reqTim);");
+        $stmnt->execute([
+          'curUs' => $curUser,
+          'othUs' => $othUser,
+          'friReq' => $friendAddName,
+          'reqTim' => date("y-m-d H:i:s"),
+        ]);
+        echo "success";
+      }
       break;
     case "accept":
       /* Gets activated when someone accepts a friend request */
       $requester = str_replace("'", "", $_POST['requester']);
       $accepter = str_replace("'", "", $_POST['accepter']);
 
-      $stmnt = $container->database->prepare("INSERT INTO friends (account_id, account_friended, friends_since)
-                                              VALUES(:accId, :accFri, :friSinc)");
-      $stmnt->execute([
-        'accId' => $accepter,
-        'accFri' => $requester,
-        'friSinc' => date("y-m-d H:i:s")
-      ]);
+      $friendAddName = "unset";
 
-      $stmnt = $container->database->prepare("DELETE FROM friend_request WHERE current_user_id = :cuid AND other_user_id = :otuid AND friend_requester = :frireq");
-      $stmnt->execute([
-        'cuid' => $requester,
-        'otuid' => $accepter,
-        'frireq' => $requester,
-      ]);
+      $stmnt = $container->database->prepare("SELECT account_name, account_id FROM accounts WHERE account_id = :accId");
+      $stmnt->execute(['accId' => $requester]);
+      $result = $stmnt->fetchAll(PDO::FETCH_ASSOC);
+      if(count($result) > 0){
+        foreach($result as $data){
+          $friendAddName = $data['account_name'];
+        }
+      }
+      if($friendAddName != "unset"){
+        $stmnt = $container->database->prepare("INSERT INTO friends (account_id, account_friended, friends_since)
+        VALUES(:accId, :accFri, :friSinc)");
+        $stmnt->execute([
+          'accId' => $accepter,
+          'accFri' => $requester,
+          'friSinc' => date("y-m-d H:i:s")
+        ]);
 
-      echo "success";
+        $stmnt = $container->database->prepare("DELETE FROM friend_request WHERE current_user_id = :cuid AND other_user_id = :otuid AND friend_requester = :frireq");
+        $stmnt->execute([
+          'cuid' => $requester,
+          'otuid' => $accepter,
+          'frireq' => $friendAddName,
+        ]);
+
+        echo "success";
+      }
       break;
     case "check":
       /* Gets activated when someone goes into their friend tab */
@@ -53,7 +79,7 @@ if(isset($_POST['type']) && $_POST['type'] != ""){
               if($_POST['userId'] != $data['friend_requester']){
                 echo "<div class='row friend-request'>";
                 echo "<p>".$data['friend_requester']."</p>";
-                $array = array('requester' => $data['friend_requester'],
+                $array = array('requester' => $data['current_user_id'],
                  'accepter' => str_replace("'", "", $_POST['userId']));
                 $functionArgs = json_encode($array);
                 echo "<button type='button' onclick='AcceptRequest(".$functionArgs.")' class='btn-flat white-text blue waves-effect waves-light'>"."<i class='material-icons'>done</i>"."</button>";
